Moving Miro transforming tests
