+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/transformer/MiroTransformerFeatureTest.scala
@@ -4,7 +4,8 @@ import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.models.aws.SQSMessage
 import uk.ac.wellcome.models.Work
 import uk.ac.wellcome.models.transformable.miro.MiroTransformable
-import uk.ac.wellcome.test.utils.{MessageInfo, MiroTransformableWrapper}
+import uk.ac.wellcome.test.utils.MessageInfo
+import uk.ac.wellcome.transformer.transformers.MiroTransformableTransformerWrapper
 import uk.ac.wellcome.transformer.utils.TransformerFeatureTest
 import uk.ac.wellcome.utils.JsonUtil
 
@@ -12,7 +13,7 @@ class MiroTransformerFeatureTest
     extends FunSpec
     with TransformerFeatureTest
     with Matchers
-    with MiroTransformableWrapper {
+    with MiroTransformableTransformerWrapper {
 
   val queueUrl: String = createQueueAndReturnUrl("test_miro_transformer")
   override val flags: Map[String, String] = Map(
